Version 1.0: Added error callback etc
This has been stable for a long while, but was not pushed to github.
This version has an error callback to track parse error, plus various
other utilities, like a way to provide function to allocate the
(optional stack).

Also, the format was extended a little to deal with single quoted
strings.

Signed-off-by: Michel Pollet <buserror@gmail.com>